build system: fix compiler warnings

+++ scripts/kconfig/mconf.c
@@ -482,7 +482,8 @@ static int exec_conf(void)
 
 	*argptr++ = NULL;
 
-	pipe(pipefd);
+	if (pipe(pipefd))
+		_exit(EXIT_FAILURE);
 	pid = fork();
 	if (pid == 0) {
 		sigprocmask(SIG_SETMASK, &osset, NULL);
@@ -846,9 +847,11 @@ static void conf(struct menu *menu)
 static void show_textbox(const char *title, const char *text, int r, int c)
 {
 	int fd;
+	int len = strlen(text);
 
 	fd = creat(".help.tmp", 0777);
-	write(fd, text, strlen(text));
+	if (write(fd, text, len) != len)
+		exit(1);
 	close(fd);
 	show_file(".help.tmp", title, r, c);
 	unlink(".help.tmp");
